cast longCells to double when calculating delta between cells

+++ src/model.h
@@ -18,13 +18,13 @@ protected:
     size_t _currentStep{0};
     Planet _currentPlanet;
     std::vector<Planet> _computedPlanets;
-    std::vector<std::map<std::string, float>> _allAtmos;
+    std::vector<std::map<std::string, double>> _allAtmos;
     bool _verbose{false};
     std::string _outputDir;
 
 public:
     // constructor for the planet specifying the number of model steps and the planet
-    Model(size_t steps, Planet planetStart, std::vector<std::map<std::string, float>> atmos, 
+    Model(size_t steps, Planet planetStart, std::vector<std::map<std::string, double>> atmos, 
           bool verbose, std::string outputDir);
 
     // run the model
@@ -37,14 +37,14 @@ public:
     virtual void calcTemps() = 0;
 
 
-    std::map<SurfaceType, float> albedoMap = {{land, 0.2}, {sea, 0.06}, {ice, 0.6}};
+    std::map<SurfaceType, double> albedoMap = {{land, 0.2}, {sea, 0.06}, {ice, 0.6}};
 
 };
 
 class SerialModel : public Model {
 
 public:
-    SerialModel(size_t steps, Planet planetStart, std::vector<std::map<std::string, float>> atmos, 
+    SerialModel(size_t steps, Planet planetStart, std::vector<std::map<std::string, double>> atmos, 
                 bool verbose, std::string outputDir);
     
     // fill in the temperatures at one step
@@ -55,7 +55,7 @@ public:
 class AccelModel : public Model {
 
 public:
-    AccelModel(size_t steps, Planet planetStart, std::vector<std::map<std::string, float>> atmos, 
+    AccelModel(size_t steps, Planet planetStart, std::vector<std::map<std::string, double>> atmos, 
                bool verbose, std::string outputDir);
     
     // fill in the temperatures at one step
